Final Project - Test Code
Week 6 of the Crash Course in Python involved creating a "Word Cloud" inside a Jupyter notebook. I used Pycharm to test running the code to clean the text and build a dictionary with occurrence counts. This code snippet differs slightly from the final submitted code, i will place that in a separate file fore viewing.