Final submitted code block
This is the final submitted code, i used a translation table to quickly and efficiently remove punctuation, and a list comprehension to build cleaned_file_contents. I again used another list comprehension to lower case each word, and ensure we are not adding words that exist within the uninteresting_words list.

The final loop is a simple for loop to count occurrences of words to be sent to the wordcloud module. This gave the expected result, and generated a complete word cloud within the associated Jupyter notebook used for grading, the submission was successful and passed grade.