Create WorkingWithCSV
This was the final script created for the Module 2 graded assessment in the "Using Python to Interact with the Operating System" course. This week largely covered using the os module to interact with the OS, and the csv module to work with csv files. While the final graded assessment leaned more on the csv side, i still learned a ton here.