Create changeJane.py
Part of Week 6 "Using Python to interact with the OS" graded test, first script was findJane.sh, this does the actual name change as described in the test. I did however spend 3 days trying to get subprocess.run to work, but determined there might be a versioning issue with the VM google provides for this lab, as it would throw errors no matter how it was typed out. I ended up going with shutil.move and got the exact same results, and this worked.